Robust inline malware alias support
In order to prevent the need for a geometric increase in "duplicate-of"
Relationship objects between different malware objects which represent
different AV vendors' names for one particular malware sample, this
removes the :name key, in favor of an array of MalwareName nodes, which
combine the AV Vendor's name ("ClamAV", for example), the given name or
alias of the malware that the vendor uses, and optionally, the signature
name that corresponds to that name designation.

This will allow us to have robust inlining of malware name aliases
for multiple AV vendors, without requiring an n^2 growth in the number
of Relationship objects over time, and simplify our ability to curate
the other descriptive fields such as the long description and kill chain
phases.

Secondary important change is the :labels key is now optional.  Certain
sources, such as abuse.ch, do not actually give MalwareLabel data, so we
don't always have the ability to assign labels to a given malware family.

Progress toward abuse.ch SSL blacklist importer
(https://github.com/threatgrid/hydrant/issues/84)