Tutorial on Modeling Threat Intel in CTIM
It's becoming increasingly clear that we need better tutorial documentation
for client developers and data analysts on how to encode threat intelligence
data in CTIM.

Accordingly, this tutorial will cover the essential structure of the most
common CTIM entity types, as well as how to package bundles, best practices on
choosing external and transient IDs, recommendations on valid times, relationship
types, and a tutorial on how to test bundles in a local CTIA container.

Nothing in this tutorial requires access to any private Cisco APIs or software
repositories.  Nor does anything in this tutorial mention Cisco products outside
CTIA and CTIM.