Incident entity description clarifications
This update clarifies the definition and purpose of the incident entity,
specifically distinguishing the concepts of "incident" and "event", as
defined in NIST Special Publication 800-61, Revision 2, on Computer Security
Incident Handling, per discussions with Eric Hulse. The goal in making these
distinctions explicit in our schema is to clarify future model mappings for
complex adverse events using the `sighting` entity, and to give guidance for
continued development of the `incident` entity, which is explicitly for the
purpose of modeling *incident response* data.

Also includes field descriptions for `IncidentTime` inline fields, to help
clarify their intent and purpose. Additional fields forthcoming.