Improve swagger docs (#449)
https://github.com/advthreat/iroh/issues/8946

Via flanders changes, adds examples to the json docs and swagger metadata.

We reuse `:default` to populate examples here. This has undesirable effects on the documentation. I reviewed the existing documentation around defaults, and many were not appropriate for user-facing docs. The closest to useful was that tlp defaults to green, but this is configurable in CTIA so that is debatable.

Instead of developing another flanders feature, I disabled default values from printing in the docs. We might want to revisit this if we actually use defaults in a user-facing way (not just generating examples).