XDR-15964 add detection_sources to incident (#457)
Epic: https://cisco-sbg.atlassian.net/browse/XDR-15686
Closes: https://cisco-sbg.atlassian.net/browse/XDR-15694

Summary: Adds optional `detection_sources` field to the Incident object. `detection_sources` differs from `source` in that `source` is the promoter of the incident, whereas `detection_sources` is the a set of aggregated sources of detections, such as sightings, that are found in the incident. This is meant be a searchable field on the Incidents List in the UI (see the epic above) and as such must be available via Lucene query. 

Local generative and testing steps:
1. Ran `lein doc` and commited results
2. Ran `lein test`, which passed