Version 1.3.26
I accidentally rebase-merged the release branch for 1.3.25, so the Clojars entry
links to an orphaned commit. While it is a matter of metadata, there
are obviously good reasons to want the github commit link to be accurate. We don't
have a great way to recover from this, but to mitigate the issue we decided to
immediately release a new version so at least the latest version doesn't have
this problem.